backend/requirements.txt
fastapi[stable]
uvicorn
sqlalchemy
psycopg2-binary
pymongo
redis
passlib[bcrypt]
pyjwt
